Refactor search adapters and add acceptance checks
